implemented test for switching between files
